design : caption 내부 비율 조절 #109
