+++ fe/src/components/daily-recommendation/DailyRecBar.tsx
@@ -16,7 +16,7 @@ export const DailyRecBar = ({ daily, dailyChartData }: Props) => {
   return (
     <div
       className={cn(
-        'flex items-center justify-between rounded-md px-4 py-3',
+        'flex h-full items-center justify-between rounded-md px-4 py-3',
         'bg-brand-subtle dark:bg-muted/40',
         'dark:border-border border border-transparent',
         'border-l-brand border-l-4'
